src/pytorch/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
fastapi==0.86.0
h11==0.14.0
idna==3.4
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pydantic==1.10.2
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wrapt==1.14.1
python-multipart
